WIP: Start work on a Render Engine object
Issue:
The temporary window-holding-gpu-state method of ownership has reached
its limit. In order to move past it, we need to define a whole render
engine which will own the window, resources, etc..

This progress commit adds the ResourceManager struct. This is a resource
holder, which manages resource creation, destruction, based on a
reference counted handle which is returned every time we access the
resource. When the reference counter hits zero, then the resource will
be marked for destruction based on its defined lifetime. After the
lifetime expires, the resource will be released sometime in the future
depending on how many resources are wanting to be destroyed in that
moment.

This resource manager will be used for many binary data types that need
to be loaded into memory and dropped from memory throughout the program
lifetime. For example; texture data, mesh data, audio data, map data,
...